rename things back into place, fix bugs.
