+++ src/ug_guts.c
